updated code in preparation for v4.1.0
- new Mnet-test script, testing framework for user automation scripts
    allows for regression testing of multiple Mnet script scenarions
    handles record/replay of Mnet::Expect, Mnet::SNMP, and Mnet::Ping sessions
    refer to `perldoc Mnet::script::Mnet-test` for more information

- added Mnet::Expect::IOS to distribution, handling many cisco specific tasks
    push enable config commands, and optional backout commands on errors
    copy and delete files, immediate and scheduled reloads, ping, and write mem
    parse show ver, run and int commands, parse and modify config stanzas

- Mnet::Expect --expect-record and --expect-replay work for multiple sessions
    command line record/replay handles multiple device connections from script

- Mnet::Expect will now gracefully handle missing expect-command substitutions
    shell user is used if --expect-username missing for default ssh attempts
    notification issued if other expect-command substitutions are missing

- Mnet::Expect new --expect-login forces login without username or password
    otherwise would connect directly to spawned expect-command without login

- Mnet perl tests will fail if standard perl or Mnet modules are missing

- minor documentation fixes and clarification updates

+++ lib/Mnet/SNMP.pm
@@ -19,15 +19,14 @@ already configured.
 Usage examples:
 
  use Mnet;
- use SNMP;
+ use Mnet::SNMP;
  my $cfg = &object({
      'object-address' => 'hub1-rtr',
      'snmp-community' => 'secret',
  });
  my $oid1 = '.1.3.6.1.2.1.1.5.0';
  my $value = &snmp_get($oid1);
- my $value = &snmp_get($oid1, {'snmp-retries' => 1});
- my %table = {};
+ my %table = ();
  my @oids = ('.1.3.6.1.2.1.1.1.0', '.1.3.6.1.2.1.1.5.0');
  &snmp_bulkget(\@oids, \%table);
  my $oid = '.1.3.6.1.2.1.2.2.1';
@@ -129,12 +128,10 @@ BEGIN {
         and not $cfg->{'snmp-replay'};
 
     # remove any old snmp-record file
-    if ($cfg->{'snmp-record'}) {
+    if ($cfg->{'snmp-record'} and -e $cfg->{'snmp-record'}) {
         &dbg("deleting snmp-record file $cfg->{'snmp-record'}");
         unlink($cfg->{'snmp-record'})
-            or croak "snmp-record $cfg->{'snmp-record'} del err $!"
-            if -e $cfg->{'snmp-record'};
-        &dtl("old snmp-record $cfg->{'snmp-record'} deleted");
+            or croak "snmp-record $cfg->{'snmp-record'} del err $!";
     }
 
 # end of module initialization
@@ -473,9 +470,8 @@ the query result.
 
     # log and optionally record snmp value
     } else {
-        &dbg("$log snmp-record oid $oid value = '$result->{$oid}'");
+        &dbg("$log oid $oid value = '$result->{$oid}'");
         if ($args->{'snmp-record'}) {
-            &dbg("$log snmp-record oid $oid value = '$result->{$oid}'");
             &snmp_record($args->{'snmp-record'}, $oid, $result->{$oid});
         }
     }
